requirements.txt
appdata==2.1.2
asgiref==3.7.2
click==8.0.3
colored==1.4.3
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
to-requirements.txt==1.3.0
typing_extensions==4.7.1
